+++ packages/fetch-router/src/index.ts
@@ -13,33 +13,33 @@ export type { RequestHandler, BuildRequestHandler } from './lib/request-handler.
 export type { Controller, Action, BuildAction } from './lib/controller.ts'
 
 export {
-  createDestroy,
-  createDestroy as destroy, // shorthand
-  createGet,
-  createGet as get, // shorthand
-  createHead,
-  createHead as head, // shorthand
-  createOptions,
-  createOptions as options, // shorthand
-  createPatch,
-  createPatch as patch, // shorthand
-  createPost,
-  createPost as post, // shorthand
-  createPut,
-  createPut as put, // shorthand
-} from './lib/route-helpers.ts'
-
-export {
   Route,
   createRoutes,
   createRoutes as route, // shorthand
 } from './lib/route-map.ts'
-export type { RouteMap, RouteDefs, RouteDef } from './lib/route-map.ts'
+export type { BuildRoute, RouteMap, RouteDefs, RouteDef } from './lib/route-map.ts'
 
 export { createRouter } from './lib/router.ts'
 export type { Router, RouterOptions } from './lib/router.ts'
 
 // Route helpers
+
+export {
+  createDeleteRoute,
+  createDeleteRoute as del, // shorthand
+  createGetRoute,
+  createGetRoute as get, // shorthand
+  createHeadRoute,
+  createHeadRoute as head, // shorthand
+  createOptionsRoute,
+  createOptionsRoute as options, // shorthand
+  createPatchRoute,
+  createPatchRoute as patch, // shorthand
+  createPostRoute,
+  createPostRoute as post, // shorthand
+  createPutRoute,
+  createPutRoute as put, // shorthand
+} from './lib/route-helpers/method.ts'
 
 export {
   createFormRoutes,